rename pages
