+++ mobile/lib/main.dart
@@ -1,5 +1,5 @@
-import 'package:flutter/material.dart';
-import 'package:weeb_dev_my_movie_list/pages/home.dart';
+import 'package:flutter/material.dart'; 
+import 'package:weeb_dev_my_movie_list/pages/home_page.dart';
 import 'package:weeb_dev_my_movie_list/util/helpers/init_helper.dart'; 
 import 'package:weeb_dev_my_movie_list/util/style/themes.dart'; 
 
@@ -26,7 +26,7 @@ class _MyMovieListAppState extends State<MyMovieListApp> {
   Widget build(BuildContext context) {
     return MaterialApp( 
       theme: Themes.mainTheme,
-      home: Home()
+      home: HomePage()
     );
   }
 }
